Try next address if socket() fails with EAFNOSUPPORT or EPROTONOSUPPORT.
getaddrinfo() may return both IPv6 and IPv4 addresses, especially for
localhost.  If IPv6 is disabled on the machine, and socket() requests
an IPv6 socket, then it fails.  In this case, we should try next
address instead of retuning error right away.

Previously, make test would fail depending on the machine.