Remove memcached_last_error checking in memcached function testing.
In Naver eco-square VM environment that I used for testing,
both IPv6 and IPv4 addresses are enabled.
At first, the client uses IPv6 to connect to a cache node.
But, the connect request fails with "connection refused" error.
And, the next connect request with IPv4 get the success response.

According to IP settings of test machine,
memcached_last_error might be set with some errors.
To make "make test" to succeed in any environment,
we remove memcached_last_error checking.